adding voided trail

+++ db/migrate/20220831190729_create_classrooms.rb
@@ -3,7 +3,7 @@ class CreateClassrooms < ActiveRecord::Migration[5.2]
     create_table :classrooms do |t|
       t.string :name
       t.string :description
-      t.float :space
+      t.integer :space
       t.boolean :voided, default: false
       t.string :void_reason
       t.datetime :date_voided, null: true
